Right now it is impossible to get the KEYPRESS, KEYDOWN or RESIZE event before the dialog gets it. This leads to tons of really hacky code to handle things like using the ENTER key to enter data from a textbox or autocomplete.
I've talked to emil, dpup, and various other clusure people, and I think we
all agree this is the right solution. So, well, hell, I'm going to propose
that we do it...mostly because I need the change in order to fix a bunch
of bugs I have with autocomplete in the dialog.

in google, let alone the world! Eeks, I don't know where to begin!

R=eae
DELTA=12  (0 added, 5 deleted, 7 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=767005


git-svn-id: http://closure-library.googlecode.com/svn/trunk@391 0b95b8e8-c90f-11de-9d4f-f947ee5921c8